very large refactoring
