+++ pastax/__init__.py
@@ -5,11 +5,6 @@ This package provides a set of tools for building and evaluating
 
 
 from ._version import version as __version__
-from .evaluation import __all__ as evaluation_all
-from .grid import __all__ as grid_all
-from .simulator import __all__ as simulator_all
-from .trajectory import __all__ as trajectory_all
-from .utils import __all__ as utils_all
 
 
-__all__ = ["__version__", ], evaluation_all + grid_all + simulator_all + trajectory_all + utils_all
+__all__ = ["__version__", ]
