+++ pastax/dynamics/__init__.py
@@ -1,0 +1,9 @@
+"""
+This module provides dynamics examples to be used with [`pastax.simulator.BaseSimulator`][].
+"""
+
+from ._linear_uv import linear_uv, LinearUV
+from ._smagorinsky_diffusion import SmagorinskyDiffusion
+
+
+__all__ = ["linear_uv", "LinearUV", "SmagorinskyDiffusion"]
