+++ pastax/trajectory/_trajectory_ensemble.py
@@ -1,18 +1,17 @@
 from __future__ import annotations
-from typing import Callable, ClassVar, Dict
+from typing import ClassVar, Dict
 
-from jaxtyping import Array, ArrayLike, Float, Int
+from jaxtyping import Array, Float, Int
 from matplotlib.collections import LineCollection
 import matplotlib.pyplot as plt
 import numpy as np
 import xarray as xr
 
-from ..utils.unit import UNIT, Unit, units_to_str
+from ..utils._unit import UNIT, Unit, units_to_str
 from ._state import State
-from .state import Location
+from ._states import Location
 from ._timeseries_ensemble import TimeseriesEnsemble
-from .trajectory import Timeseries, Trajectory
-from ._unitful import Unitful
+from ._trajectory import Trajectory
 
 
 class TrajectoryEnsemble(TimeseriesEnsemble):
@@ -54,12 +53,10 @@ class TrajectoryEnsemble(TimeseriesEnsemble):
         Computes the separation distance for each ensemble trajectory.
     steps()
         Returns the steps of the trajectories.
+    to_xarray()
+        Converts the [`pastax.trajectory.TrajectoryEnsemble`][] to a `xarray.Dataset`.
     from_array(values, times, unit=UNIT["°"]
-        Creates a TrajectoryEnsemble from an array of values and time points.
-    to_dataarray()
-        Converts the trajectory ensemble locations to a dict of xarray DataArray.
-    to_dataset()
-        Converts the trajectory ensemble to a xarray Dataset.
+        Creates a [`pastax.trajectory.TrajectoryEnsemble`][] from arrays of values and time points.
     """
 
     members: Trajectory
@@ -124,28 +121,6 @@ class TrajectoryEnsemble(TimeseriesEnsemble):
             The origin of the trajectories.
         """
         return self.members.origin
-
-    def crps(
-        self,
-        other: Trajectory,
-        distance_func: Callable[[Trajectory, Trajectory], Unitful | ArrayLike] = Trajectory.separation_distance
-    ) -> Timeseries:
-        """
-        Computes the Continuous Ranked Probability Score (CRPS) for the ensemble.
-
-        Parameters
-        ----------
-        other : Trajectory
-            The other trajectory to compare against.
-        distance_func : Callable[[Trajectory, Trajectory], Unitful | ArrayLike], optional
-            The distance function to use, defaults to Trajectory.separation_distance.
-
-        Returns
-        -------
-        Timeseries
-            The CRPS for the ensemble.
-        """
-        return super().crps(other, distance_func)
 
     def liu_index(self, other: Trajectory) -> TimeseriesEnsemble:
         """
@@ -288,6 +263,17 @@ class TrajectoryEnsemble(TimeseriesEnsemble):
         """
         return self.map(lambda trajectory: trajectory.steps())
 
+    def to_xarray(self) -> xr.Dataset:
+        """
+        Converts the [`pastax.trajectory.TrajectoryEnsemble`][] to a `xarray.Dataset`.
+
+        Returns
+        -------
+        xr.Dataset
+            The corresponding `xarray.Dataset`.
+        """
+        return xr.Dataset(self._to_dataarray())
+
     @classmethod
     def from_array(
         cls,
@@ -298,69 +284,58 @@ class TrajectoryEnsemble(TimeseriesEnsemble):
         **_: Dict
     ) -> TrajectoryEnsemble:
         """
-        Creates a TrajectoryEnsemble from an array of values and time points.
+        Creates a [`pastax.trajectory.TrajectoryEnsemble`][] from arrays of values and time points.
 
         Parameters
         ----------
         values : Float[Array, "member time 2"]
-            The values for the members of the trajectory ensemble.
+            The array of (latitudes, longitudes) values for the members of the trajectory ensemble.
         times : Float[Array, "time"]
-            The time points for the timeseries.
+            The time points for the trajectories.
         unit : Unit | Dict[Unit, int], optional
-            Unit of the trajectory locations, defaults to UNIT["°"].
+            Unit of the trajectories locations, defaults to UNIT["°"].
         id : Int[Array, ""], optional
-            The ID of the trajectory, defaults to None.
+            The ID of the trajectories, defaults to None.
 
         Returns
         -------
         TrajectoryEnsemble
-            The TrajectoryEnsemble created from the array of values and time points.
+            The corresponding [`pastax.trajectory.TrajectoryEnsemble`][].
         """
         return super().from_array(values, times, unit=unit, id=id)
 
-    def to_dataarray(self) -> Dict[str, xr.DataArray]:
+    def _to_dataarray(self) -> Dict[str, xr.DataArray]:
         """
-        Converts the ensemble states to a dictionary of xarray DataArrays.
+        Converts the [`pastax.trajectory.TrajectoryEnsemble`][] to a dictionary of `xarray.DataArray`.
 
         Returns
         -------
         Dict[str, xr.DataArray]
-            A dictionary where keys are the variable names and values are the corresponding xarray DataArrays.
+            A dictionary where keys are the variable names and values are the corresponding `xarray.DataArray`.
         """
         member = np.arange(self.size)
         times = self.members.times.to_datetime()
         unit = units_to_str(self.unit)
 
-        id_da = xr.DataArray(
+        wmo_da = xr.DataArray(
             data=self.id,
-            dims=["member"],
-            coords={"member": member},
-            name="drifter id"
+            dims=["traj", "obs"],
+            coords={"id": ("traj", member)},
+            name="WMO"
         )
         latitude_da = xr.DataArray(
             data=self.latitudes,
-            dims=["member", "time"],
-            coords={"member": member, "time": times},
-            name="latitude",
+            dims=["traj", "obs"],
+            coords={"id": ("traj", member), "time": ("obs", times)},
+            name="lat",
             attrs={"units": unit}
         )
         longitude_da = xr.DataArray(
             data=self.longitudes,
-            dims=["member", "time"],
-            coords={"member": member, "time": times},
-            name="longitude",
+            dims=["traj", "obs"],
+            coords={"id": ("traj", member), "time": ("obs", times)},
+            name="lon",
             attrs={"units": unit}
         )
 
-        return {"id": id_da, "latitude": latitude_da, "longitude": longitude_da}
-
-    def to_dataset(self) -> xr.Dataset:
-        """
-        Converts the ensemble states to an xarray Dataset.
-
-        Returns
-        -------
-        xr.Dataset
-            An xarray Dataset containing the ensemble states.
-        """
-        return xr.Dataset(self.to_dataarray())
+        return {"WMO": wmo_da, "latitude": latitude_da, "longitude": longitude_da}
